requirements.txt
openai
python-dotenv
docling